requirements.txt
loguru
python-dotenv
python-telegram-bot>=20.7
python-telegram-bot[job-queue]
gspread
gspread-asyncio
pandas
yamager